requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.2
django-bootstrap-v5==1.0.11
djangorestframework==3.15.2
psycopg==3.2.3
psycopg-binary==3.2.3
python-dotenv==1.0.1
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
